requirements.txt
# Web Dashboard
streamlit

# Email & Data Processing
beautifulsoup4
lxml
pandas

# Security & Environment
python-dotenv

# Networking (Optional but recommended for stability)
urllib3